backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.3
python-dotenv==1.0.1
